Refactor activity visibility logic and remove Audience trait
Moved the activity visibility determination logic from the Audience trait to a new get_activity_visibility() function in functions.php. Updated controllers to use the new function and removed the now-unnecessary trait and its tests. Added comprehensive tests for get_activity_visibility in the functions test suite.
